+++ Changelog.md
@@ -6,6 +6,8 @@
    for the text-to-speech feature.
  * The `speak` section of the yaml file can have subsections with differing
    languge settings, to allow multi-language files.
+ * The text-to-speech feature will use either pico2wave or speak, depending on
+   what is available.
 
 ## 1.0 (2015-01-29)